Replaced param and and added line breaks in docs
The @param denoting updateRole had to be changed to requester to match
other documentation. This should be changed in the future to a more
approprate name. Line breaks were also added where psudo-bullets exist
in java docs. This prevents malformed docs when running the Google Style
Formatter, which likes place the bullets on one line.